Stop using pod structure. Move styles to addon
