+++ app/components/tf-button.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-ticketfly-buttons/components/tf-button/component';
+export { default } from 'ember-ticketfly-buttons/components/tf-button';
